broke up the charting logic. single file was getting too big
